requirements.txt
flask
pdfminer.six
spacy
pytextrank
